memory-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-mcp"
version = "0.2.0"
description = "MCP server for AI long-term memory - Let AI remember across sessions!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
migrate = [
    "chromadb>=0.5.0",
]

[project.scripts]
memory-mcp = "memory_mcp.server:main"
memory-migrate = "memory_mcp.migrate:main"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_mcp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
per-file-ignores = {"src/memory_mcp/server.py" = ["E501"]}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
